+++ app/dashboard/drone-dashboard/src/pages/Overview.js
@@ -5,7 +5,11 @@ import axios from 'axios';
 import CommandSender from '../components/CommandSender';
 import DroneWidget from '../components/DroneWidget';
 import ExpandedDronePortal from '../components/ExpandedDronePortal';
-import { attachDroneRuntimeClock, normalizeTelemetryResponse } from '../constants/fieldMappings';
+import {
+  attachDroneRuntimeClock,
+  extractServerNowMs,
+  normalizeTelemetryResponse,
+} from '../constants/fieldMappings';
 import { normalizeComparableId } from '../utilities/missionIdentityUtils';
 import { getBackendURL, getTelemetryURL } from '../utilities/utilities';
 import '../styles/Overview.css';
@@ -58,12 +62,16 @@ const Overview = ({ setSelectedDrone }) => {
     const fetchData = async () => {
       try {
         const response = await axios.get(url);
-        const normalizedTelemetry = normalizeTelemetryResponse(response.data || {});
+        const clockMeta = {
+          receivedAtMs: Date.now(),
+          serverNowMs: extractServerNowMs(response.headers),
+        };
+        const normalizedTelemetry = normalizeTelemetryResponse(response.data || {}, clockMeta);
         const dronesArray = Object.keys(normalizedTelemetry).map((hw_ID) => attachDroneRuntimeClock({
           ...(configByHwId[normalizeComparableId(hw_ID)] || {}),
           hw_ID,
           ...normalizedTelemetry[hw_ID],
-        }));
+        }, clockMeta));
 
         const validDrones = dronesArray.filter(
           (drone) =>